+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Futoshiki solver".
 - Tests for "Frog exchange".
 - Tests for "Fair numbering".
+- Tests for "A coin guessing game".
 
 ## [1.11.0] - 2022-07-31
 ### Added